Updated about half the methods of the DB class to accommodate security fixes

+++ db.py
@@ -1,10 +1,23 @@
-'''
+"""
 AUTHOR: COBY JOHNSON
-PROJECT: GENERIC --> DATABASE CLASS
-LAST UPDATE: 2/17/2014
+PROJECT: SQLite3-DB
+LAST UPDATE: 3/3/2014
+
+DONE:
++ DB.init
++ DB.dictToStrings
++ DB.createTable
++ DB.dropTable
++ DB.closeDB
++ DB.printTable
++ DB.getConstraints
++ DB.error
++ DB.warning
 
 TODO:
 -!!!NEED TO MAKE ALL FUNCTIONS SAFE FROM SQL INJECTION ATTACKS!!!-
+
+-REVISIT ALL FUNCTIONS NOT IN DONE LIST AND FIX TO USE DICTIONARY PARAMETERS
 
 -PROPERLY TEST INNER JOIN FUNCTION TO PROVE FUNCTIONALITY
     -WRITE COMMENTS ON PARAMETERS
@@ -24,34 +37,58 @@ EVENTUALLY DO SOMETIME IN THE FAR FUTURE:
 -STOP HACKERS BY SCARING THE CRAP OUT OF THEM IF THEY TRY INTENTIALLY MALICOUS CODE
     -RETURN URL PAGE WITH THEIR HOME ADDRESS AND SAYING WE HAVE NOTIFED THE LOCAL FBI
      OF THEIR ILLEGAL ACTIVITIES.
-'''
+"""
 
 import sqlite3 as sql
 
 class DB:
     def __init__(self, name):
-        #LOAD DATABASE
-        self.db = sql.connect(name, detect_types=sql.PARSE_DECLTYPES|sql.PARSE_COLNAMES)
-        #CREATE CURSOR
+        #Load DB
+        self.db = sql.connect(name)
+        #Create DB cursor
         self.cursor = self.db.cursor()
 
     #createTable(self,
-    #            table, #TABLE NAME
-    #            info)  #(COLUMN NAME 1, ..., COLUMN NAME X)
+    #            table, #Table name
+    #            info)  #(Column_name_1, ..., Column_name_X)
     def createTable(self, table, info):
-        self.cursor.execute('''CREATE TABLE {0}{1}'''.format(table, info))
-        self.db.commit()
-        return
+        try:
+            #print '''CREATE TABLE {0} {1}'''.format(table, info)
+            self.cursor.execute('''CREATE TABLE {0} {1}'''.format(table, info))
+            self.db.commit()
+            return
+        except:
+            self.error('Table "{0}" already exists.'.format(table))
         
     #dropTable(self,
-    #          table) #TABLE NAME
+    #          table) #Table name
     def dropTable(self, table):
-        self.cursor.execute('''DROP TABLE IF EXISTS {0}'''.format(table))
-        self.db.commit()
+        answer = 'n'
+        self.warning('This will delete table "{0}" with all of it data'.format(table))
+        answer = raw_input('Are you sure you want to drop table "{0}"? (y/N)'.format(table))
+        if (answer.lower == 'y'):
+            #print '''DROP TABLE IF EXISTS {0}'''.format(table)
+            self.cursor.execute('''DROP TABLE IF EXISTS {0}'''.format(table))
+            self.db.commit()
+        else:
+            print 'Disaster averted. Table "{0}" was not changed.'.format(table)
         return
 
+    #clearTable(self,
+    #           table) #Table name
+    def clearTable(self, table):
+        answer = 'n'
+        self.warning('This will delete all data from table "{0}"'.format(table))
+        answer = raw_input('Are you sure you want to delete all records from table "{0}"? (y/N)'.format(table))
+        if (answer.lower() == 'y'):
+            self.cursor.execute('''DELETE FROM {0}'''.format(table))
+            self.db.commit()
+            print 'Table "{0}" successfully cleared.'.format(table)
+        else:
+            print 'Disaster averted. Table "{0}" was not changed.'.format(table)
+
     #alterTable(self,
-    #           table,   #TABLE NAME
+    #           table,   #Table
     #           command) #SUPPORTS ADD column, RENAME TO table
     ### NOTE: SQLITE3 DOES NOT SUPPORT: RENAME column, DROP column ###
     def alterTable(self, table, command):
@@ -62,39 +99,18 @@ class DB:
 
     #insertRow(self,
     #          row,  #ROW NAME
-    #          info) #(INFO 1, ..., INFO X)
+    #          info) #{key0:value0, ..., keyX:valueX}
     def insertRow(self, row, info):
+        (keys, values) = self.paramDict(info)
+        #print '''INSERT INTO {0} ({1}) VALUES ({2})'''.format(row, keys, values)
         try:
-            keys = ""
-            values = ""
-            for k in info.keys():
-                keys += k + ", "
-                values += ":" + k + ", "
-            keys = keys[:len(keys)-2]
-            values = values[:len(values)-2]
-            
-            print '''INSERT INTO {0} ({1}) VALUES ({2})'''.format(row, keys, values)
+            #Will add row to DB unless there is a duplicate PRIMARY/UNIQUE key constraint.
             self.cursor.execute('''INSERT INTO {0} ({1}) VALUES ({2})'''.format(row, keys, values), info)
-            
-            #print '''INSERT INTO {0} VALUES{1}'''.format(row, info)
-            #self.cursor.execute('''INSERT INTO {0} VALUES {1}'''.format(row, info))
             self.db.commit()
             return
         except sql.IntegrityError:
-            ### ||| NOTE: MODIFY THIS EXCEPTION FOR YOUR DB SPECIFICALLY ||| ###
-            ### \\\                                                      /// ###
-            print '##### Duplicate value in table with unique restraint. #####'
-            temp = info.split()
-            current = self.getRow(row, 'name={0}'.format(temp[0].strip("(,)")))
-            response = raw_input('Do you want to merge --> {0},\n    with this record --> {1}\n (Y/n)?'.format(info, current))
-            if (response == 'N' or response == 'n'):
-                print 'Record discarded'
-            else:
-                num = int(temp[2].strip(','))
-                self.updateRow('MTG', 'count', current[2] + num, current[3])
-                self.db.commit()
-            ### ///                                                      \\\ ###
-            ### ||| NOTE: MODIFY THIS EXCEPTION FOR YOUR DB SPECIFICALLY ||| ###
+            #Tell user to inherit and overwrite this method
+            raise NotImplementedError('Could not insert row: You need to implement how you want to merge your ADT.')
         
     #deleteRow(self,
     #          row,       #ROW NAME
@@ -144,7 +160,6 @@ class DB:
     #          table2,    #
     #          column2,   #
     #          condition) #
-    #
     def innerJoin(self, table1, column1, table2, column2, condition):
         print '''SELECT {0}.{1}, {2}.{3} FROM {0} JOIN {2} ON {4}'''.format(table1, column1, table2, column2, condition)
         self.cursor.execute('''SELECT {0}.{1}, {2}.{3} FROM {0} JOIN {2} ON {4}'''.format(table1, column1, table2, column2, condition))
@@ -152,7 +167,7 @@ class DB:
         return results
 
     #printTable(self,
-    #           table) #TABLE NAME
+    #           table) #Table name
     def printTable(self, table):
         cursor = self.cursor.execute('''SELECT * FROM {0}'''.format(table))
         for row in cursor:
@@ -160,8 +175,11 @@ class DB:
         return
 
     #getColumnNames(self,
-    #               table) #TABLE NAME
+    #               table) #Table name
     def getColumnNames(self, table):
+        """
+        [(Column Number, Column Name, Data Type, Able to be Null/Not Null, Default Value, Is Primary key?), ..., ()]
+        """
         try:
             self.db.row_factory = sql.Row
             cursor = self.db.execute('select * from {0}'.format(table))
@@ -170,16 +188,105 @@ class DB:
             return tuple(names)
         except:
             print '##### No columns in table "{0}". #####'.format(table)
-            return ()
+            return []
+
+    #paramDict(self,
+    #          info)    #Dictionary to parse
+    """
+    Parameterizes a dictionary into appropriate string values
+    """
+    def paramDict(self, info):
+        keys = ""
+        values = ""
+        for k in info.keys():
+            keys += k + ", "
+            values += ":" + k + ", "
+        keys = keys[:len(keys)-2]
+        values = values[:len(values)-2]
+        #strings look like this
+        #   "key0, ..., keyX"
+        #   ":value0, ..., :valueX"
+        return (keys, values)
+
+    #getConstraints(self,
+    #               table) #Table name
+    """
+    Returns all the table columns that are under a unique/primary key restraint
+    Returns in the format [(table_name, [column info]), ..., (table_name, [column info])]
+    """
+    def getConstraints(self, table="ALL"):
+        
+        #Find all constraints in all tables
+        if (table == "ALL"):
+            #Get all table headers from DB
+            self.cursor.execute("select * from sqlite_master")
+            schema = self.cursor.fetchall()
+            #Pull out user created tables
+            T = []
+            for item in schema:
+                if (item[0] == 'table' and item[1] != 'sqlite_sequence'):
+                    T.append(item)
+            #Pull out unique constraints for each table
+            U = []
+            for item in T:
+                U.append((str(item[1]), self.getConstraints(item[1])))
+            return U
+
+        #Find constraints in a specific table
+        else:
+            #Get table header from DB
+            d = {}
+            d['name'] = table
+            self.cursor.execute("select * from sqlite_master where name=(:name)", d)
+            schema = self.cursor.fetchone()
+            #Does the table exist?
+            if (schema == None):
+                self.error('Failed to find constraints: Table does not exist.')
+            #Find the paranthesis
+            lp = schema[4].find('(') + 1
+            rp = schema[4].rfind(')')
+            #Type cast from unicode to string
+            schema = str(schema[4])
+            #Find table name before next step
+            name = schema[:lp-1]
+            temp = name.lower()
+            n = temp.find('create table')
+            name = name[n+12:].strip()
+            #Splice out the columnnames
+            schema = schema[lp:rp]
+            temp = schema.split(',')
+            #Remove white space
+            columns = []
+            for item in temp:
+                columns.append((item.strip(), item.lower().strip()))
+            #Pull out unique/primary key constraints and return
+            unique = []
+            for item in columns:
+                if (item[1].find('unique') != -1 or item[1].find('primary key') != -1):
+                    unique.append(item[0])
+            return [(name, unique)]
 
     #closeDB(self)
     def closeDB(self):
-        #SAVE DATABASE
+        #Save database
         self.db.commit()
-        #CLOSE CURSOR
+        #Close cursor
         self.cursor.close()
-        #CLOSE DATABASE
+        #Close database
         self.db.close()
+        return
+
+    #error(self,
+    #      message) #Error message to be displayed
+    def error(self, message):
+        self.db.rollback()
+        self.closeDB()
+        raise Exception('### {0} ###'.format(message))
+
+    #warning(self,
+    #        message) #Warning message to be displayed
+    def warning(self, message):
+        print '=== {0}! ==='.format(message)
         return
 
     #safe(self) #Prevents SQL Injection attacks
@@ -193,27 +300,38 @@ class DB:
 def main():
     db = DB('MTG.sql')
 
-    hack = "ilovehackers@hackers.net"
-    print db.safe(hack)
-    hack = "anything' or 'x'='x'"
-    print db.safe(hack)
-    hack = "steve@unixwiz.net'"
-    print db.safe(hack)
-    hack = "x' AND email IS NULL; --"
-    print db.safe(hack)
-    hack = "x' AND 1=(SELECT COUNT(*) FROM tabname); --"
-    print db.safe(hack)
-    hack = "x' AND members.email IS NULL; --"
-    print db.safe(hack)
+##    hack = "ilovehackers@hackers.net"
+##    print db.safe(hack)
+##    hack = "anything' or 'x'='x'"
+##    print db.safe(hack)
+##    hack = "steve@unixwiz.net'"
+##    print db.safe(hack)
+##    hack = "x' AND email IS NULL; --"
+##    print db.safe(hack)
+##    hack = "x' AND 1=(SELECT COUNT(*) FROM tabname); --"
+##    print db.safe(hack)
+##    hack = "x' AND members.email IS NULL; --"
+##    print db.safe(hack)
 
     db.dropTable('MTG')
+    db.dropTable('Test')
 
-    db.createTable('MTG', '(name TEXT UNIQUE, color TEXT, count INTEGER, ID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT)')
-    db.insertRow('MTG', {'name': 'Plain', 'color': 'WH', 'count': 50, 'ID': None})
-    db.insertRow('MTG', "('Swamp', 'BK', 50, NULL)")
-    db.insertRow('MTG', "('Swamp', 'BK', 50, NULL)")
+    db.createTable('MTG', '(name TEXT UNIQUE, color TEXT, count INTEGER CHECK(count > 0), ID INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT)')
+    db.insertRow('MTG', {'name': 'Plain', 'color': 'WH', 'count': 50})
+    #db.insertRow('MTG', {'name': 'Swamp', 'color': 'BK', 'count': 50})
+    db.insertRow('MTG', {'name': 'Swamp', 'color': 'BK', 'count': 50})
 
-    db.printTable('MTG')
+    #db.createTable('Test', '(test TEXT, name INTEGER PRIMARY KEY)')
+
+    #GET THIS TO WORK WITH MULTIPLE TABLES IN DATABASE
+
+    print db.getConstraints('MTG')
+    #print db.getConstraints('Test')
+    print db.getConstraints()
+
+    print db.clearTable('MTG')
+
+    #db.printTable('MTG')
     
     db.closeDB()
